1.50.66 - Fix Build 0/0: unify Auxiliary DTOs to double with explicit (decimal)<->double casts, fix AvailableQty/PackageCount vs QtyKg, fix ProductName->ProductNameAr, fix Tuple in EF query via AsEnumerable, fix SystemSettingsService OpResult return, fix Desktop double TryParse

+++ src/DatesErp.Desktop/Views/Screens/AuxiliarySetupView.xaml.cs
@@ -62,7 +62,7 @@ public partial class AuxiliarySetupView : UserControl
         }
 
         string method = (MethodBox.SelectedItem as ComboBoxItem)?.Tag as string ?? "ByUnit";
-        decimal.TryParse(WeightBox.Text, out var w);
+        double.TryParse(WeightBox.Text, out var w);
 
         try
         {
